Move Match under errors module
Moves Match class unider errors module in order to make it easier
to use from other places, avoiding cyclic imports.
